feat(observability): port ironclaw observability stack to dasclaw_observability (ADR-152 §3 F3.5)
## 背景 / 目标

ADR-152 §3 阶段 F3.5 要求把 ironclaw 自带的 observability（traits / log / multi / noop / prompt_cache）下沉到 `crates/dasclaw_observability`，让 admin-backend 和未来 headless agent 不依赖 desktop client 也能直接调用同一套 Observer。

PR #629 之前 `dasclaw_observability` 只是 W1 占位 skeleton（一个 placeholder trait + ObservabilityEvent struct + SkeletonError）。本 PR 把 ironclaw 的真实实现搬过来作为正式骨架。

Closes #631
Refs ADR-152 §3 F3.5

## 改动范围

- `git mv desktop-client/ironclaw/src/observability/{log,multi,noop,prompt_cache,traits}.rs → crates/dasclaw_observability/src/`（5 文件 100% similarity，git 历史保留，ADR-129 verbatim）
- 各文件内 `use crate::observability::{traits, log, ...}` → `use crate::{traits, log, ...}`（仅路径重写，函数体未改）
- `crates/dasclaw_observability/src/lib.rs`：移除 W1 placeholder（`SkeletonError` / `ObservabilityEvent` / placeholder `Observer`，验证无外部使用方），改为正式 module 树 + `ObservabilityConfig` + `create_observer()` + 6 个 factory 测试，全部从 ironclaw `observability/mod.rs` verbatim 搬过来
- `crates/dasclaw_observability/Cargo.toml`：description 更新；移除占位 `thiserror`；新增 `tracing = "0.1"`
- `desktop-client/ironclaw/Cargo.toml`：新增 `dasclaw_observability = { path = "../../crates/dasclaw_observability" }`
- `desktop-client/ironclaw/src/observability/mod.rs`：缩减到 7 行 `pub use dasclaw_observability::*;` shim，所有 `crate::observability::Observer` / `LogObserver` / `MultiObserver` / `NoopObserver` / `PromptCacheMonitor` / `ObservabilityConfig` / `create_observer()` 调用路径不变

## 非目标 (What's NOT in this PR)

- 不改事件命名 schema、event 字段、log 级别（向后兼容硬约束）
- 不调整 prompt cache 命中策略或 STATS_LOG_INTERVAL
- 不动 ironclaw 任何 observer 调用点（agentic_loop 等照常用 `crate::observability::*`）
- 不在新 crate 引入 OpenTelemetry / Prometheus 后端（ADR 留作未来扩展）
- 不动 `dasclaw_observability/Cargo.toml` 的 `version = "0.0.0-w1"`（其它 ADR-114 类 grep guard 会盯版本）

## 验证

```bash
cargo check -p dasclaw_observability --tests   # Finished in 26.19s
cargo nextest run -p dasclaw_observability     # 25 passed, 0 failed
cargo check -p dasclaw --tests                 # Finished in 5m 19s
cargo fmt --all
python3.12 scripts/check_no_panics.py --base origin/xClaw
# → OK: No panic-inducing calls in changed production code.
cargo clippy --no-deps -p dasclaw_observability --all-targets -- -D warnings  # clean
cargo clippy --no-deps -p dasclaw --all-targets -- -D warnings                 # clean (6m 14s)
```

`dasclaw` 的 `multiple build targets` 警告与本 PR 无关，是预先存在的 bin 配置。

## Sources read

- AGENTS.md §强制阅读顺序、§红线、§本地管线、§PR 必填项
- docs/plans/architecture-refactor/adr-152-agent-and-capability-fusion.md §3 F3.5（本 PR 收口的 deliverable）、§4 灰度策略、§6 红线、§5 测试基线
- docs/plans/architecture-refactor/adr-129-sandbox-windows-windows-crate-adoption.md §1.3 verbatim port
- docs/plans/architecture-refactor/adr-113-hook-engine-unification.md（确认 observability 不引入第二条 hook chain，未触红线）
- docs/plans/architecture-refactor/adr-114-dasclaw-rebrand.md（确认 grep guard 不被触发）
- crates/dasclaw_observability/src/lib.rs（旧 W1 skeleton，确认 placeholder 无外部 user：`grep dasclaw_observability::(Observer|ObservabilityEvent|SkeletonError)` → 0 命中）
- desktop-client/ironclaw/src/observability/mod.rs（含 6 个 factory 测试与 ObservabilityConfig）
- desktop-client/ironclaw/src/observability/{traits,log,multi,noop,prompt_cache}.rs（确认 import 仅 `crate::observability::traits::*` 一处需重写）
- desktop-client/ironclaw/Cargo.toml lines 185-200（确认 dep 注入位置紧邻 dasclaw_lsp）
