+++ crates/dasclaw_observability/Cargo.toml
@@ -2,7 +2,7 @@
 name = "dasclaw_observability"
 version = "0.0.0-w1"
 edition = "2021"
-description = "Unified observability + rollout-trace 4 sub-tables (codex + ironclaw)."
+description = "Trait-based observability (Observer / NoopObserver / LogObserver / MultiObserver / PromptCacheMonitor) shared by ironclaw and admin-backend."
 license = "Apache-2.0 OR MIT"
 publish = false
 
@@ -10,4 +10,4 @@ publish = false
 path = "src/lib.rs"
 
 [dependencies]
-thiserror = "1"
+tracing = "0.1"
